feat: implement player statistics aggregation for match data
- Introduced a new utility function `aggregatePlayerStats` to centralize player statistics calculation from match data, improving code organization and maintainability.
- Updated the match router to utilize the new aggregation function, streamlining player data processing and enhancing performance.
- Removed redundant player statistics logic from the match router, ensuring cleaner and more efficient code.
